+++ src/cmark/node.lisp
@@ -25,15 +25,23 @@
            :documentation "Parent node of this node, or NIL if this is a root node"))
   (:documentation "A node which can have a parent node."))
 
-(defclass document-node (parent-node)
+(defclass block-node ()
+  ()
+  (:documentation "A mixin class which marks block-level nodes."))
+
+(defclass inline-node ()
+  ()
+  (:documentation "A mixin class which marks inline nodes."))
+
+(defclass document-node (parent-node block-node)
   ()
   (:documentation "Root node of the entire document."))
 
-(defclass block-quote-node (parent-node child-node)
+(defclass block-quote-node (parent-node child-node block-node)
   ()
   (:documentation "Represents a block quotation."))
 
-(defclass list-node (parent-node child-node)
+(defclass list-node (parent-node child-node block-node)
   ((type :initarg :type :accessor node-list-type :type (member :cmark-bullet-list :cmark-ordered-list)
          :documentation "The type of list")
    (delim :initarg :delim :accessor node-list-delim :type (member :period :paren)
@@ -45,89 +53,89 @@
   (:documentation
    "A list node, it can be either an unordered list (:BULLET-LIST) or an :ORDERED-LIST."))
 
-(defclass item-node (parent-node child-node)
+(defclass item-node (parent-node child-node block-node)
   ()
   (:documentation "Represents an item inside a list (bullet or unordered)."))
 
-(defclass code-block-node (child-node)
+(defclass code-block-node (child-node block-node)
   ((fence-info :initarg :fence-info :accessor node-fence-info :type string
                :documentation "Info string from a fenced code block")
    (literal :initarg :literal :accessor node-literal :type (or string null)
             :documentation "Text of the node (might be empty)."))
   (:documentation "A block of code to present verbatim."))
 
-(defclass html-block-node (child-node)
+(defclass html-block-node (child-node block-node)
   ((literal :initarg :literal :accessor node-literal :type (or string null)
             :documentation "Text of the node (might be empty)."))
   (:documentation "Block-level embedded HTML code."))
 
-(defclass custom-block-node (parent-node child-node)
+(defclass custom-block-node (parent-node child-node block-node)
   ((on-enter :initarg :on-enter :accessor node-on-enter :type string
              :documentation "Literal text to render on entering a custom node")
    (on-exit :initarg :on-exit :accessor node-on-exit :type string
             :documentation "Literal text to render on exiting a custom node"))
   (:documentation "A user-defined node which produces custom text."))
 
-(defclass paragraph-node (parent-node child-node)
+(defclass paragraph-node (parent-node child-node block-node)
   ()
   (:documentation "A paragraph of the document."))
 
-(defclass heading-node (parent-node child-node)
+(defclass heading-node (parent-node child-node block-node)
   ((level :initarg :level :accessor node-heading-level :type (integer 1 6)
           :documentation "Level of the heading"))
   (:documentation "A heading with variable level."))
 
-(defclass thematic-break-node (child-node)
+(defclass thematic-break-node (child-node block-node)
   ()
   (:documentation "Represents a thematic break within the content of the document."))
 
 ;;; inline
-(defclass text-node (child-node)
+(defclass text-node (child-node inline-node)
   ((literal :initarg :literal :accessor node-literal :type (or string null)
             :documentation "Text of the node (might be empty)."))
   (:documentation "Plain text content."))
 
-(defclass softbreak-node (child-node)
+(defclass softbreak-node (child-node inline-node)
   ()
   (:documentation "A line break which may be rendered differently depending on the rendering options."))
 
-(defclass linebreak-node (child-node)
+(defclass linebreak-node (child-node inline-node)
   ()
   (:documentation "A hard line break which does not terminate the current block level element."))
 
-(defclass code-node (child-node)
+(defclass code-node (child-node inline-node)
   ((literal :initarg :literal :accessor node-literal :type (or string null)
             :documentation "Text of the node (might be empty)."))
   (:documentation "Inline verbatim code."))
 
-(defclass html-inline-node (child-node)
+(defclass html-inline-node (child-node inline-node)
   ((literal :initarg :literal :accessor node-literal :type (or string null)
             :documentation "Text of the node (might be empty)."))
   (:documentation "Inline embedded HTML code."))
 
-(defclass custom-inline-node (parent-node child-node)
+(defclass custom-inline-node (parent-node child-node inline-node)
   ((on-enter :initarg :on-enter :accessor node-on-enter :type string
              :documentation "Literal text to render on entering a custom node")
    (on-exit :initarg :on-exit :accessor node-on-exit :type string
             :documentation "Literal text to render on exiting a custom node"))
   (:documentation "A user-defined node which produces custom text."))
 
-(defclass emph-node (parent-node child-node)
+(defclass emph-node (parent-node child-node inline-node)
   ()
   (:documentation "Container for emphatic text."))
 
-(defclass strong-node (parent-node child-node)
+(defclass strong-node (parent-node child-node inline-node)
   ()
   (:documentation "Container for strong text."))
 
-(defclass link-node (parent-node child-node)
+(defclass link-node (parent-node child-node inline-node)
   ((url :initarg :url :accessor node-url :type string
         :documentation "URL of a link node")
    (title :initarg :title :accessor node-title :type string
           :documentation "Title of the node (empty if there is no title)"))
   (:documentation "Hyperlink with target URL and title."))
 
-(defclass image-node (parent-node child-node)
+(defclass image-node (parent-node child-node inline-node)
   ((url :initarg :url :accessor node-url :type string
         :documentation "URL of an image node")
    (title :initarg :title :accessor node-title :type string
@@ -142,32 +150,18 @@
 (defgeneric inline-node-p (node)
   (:documentation
   "Whether a node is an inline node (as opposed to a block-level node).")
-  (:method ((node               node)) nil)
-  (:method ((node          text-node))   t)
-  (:method ((node     softbreak-node))   t)
-  (:method ((node     linebreak-node))   t)
-  (:method ((node          code-node))   t)
-  (:method ((node   html-inline-node))   t)
-  (:method ((node custom-inline-node))   t)
-  (:method ((node          emph-node))   t)
-  (:method ((node        strong-node))   t)
-  (:method ((node          link-node))   t)
-  (:method ((node         image-node))   t))
+  (:method ((node node))
+    nil)
+  (:method ((node inline-node))
+    t))
 
 (defgeneric block-node-p (node)
   (:documentation
   "Whether a node is a block-level node (as opposed to an inline node).")
-  (:method ((node                node)) nil)
-  (:method ((node       document-node))   t)
-  (:method ((node    block-quote-node))   t)
-  (:method ((node           list-node))   t)
-  (:method ((node           item-node))   t)
-  (:method ((node     code-block-node))   t)
-  (:method ((node     html-block-node))   t)
-  (:method ((node   custom-block-node))   t)
-  (:method ((node      paragraph-node))   t)
-  (:method ((node        heading-node))   t)
-  (:method ((node thematic-break-node))   t))
+  (:method ((node node))
+    nil)
+  (:method ((node block-node))
+    t))
 
 (defgeneric leaf-node-p (node)
   (:documentation